Add returning twitch poins, make skip command and some refactor
